move functions related to default item content to a specific file
