+++ src/app/app.routes.ts
@@ -4,7 +4,7 @@ import { BeforeUnloadGuard } from 'src/app/guards/before-unload-guard';
 import { GroupDeleteService } from './groups/data-access/group-delete.service';
 import { DefaultLayoutInitService } from './services/layout.service';
 import { urlStringFromArray } from './utils/url';
-import { appDefaultActivityRoute } from './models/routing/item-route';
+import { appDefaultActivityRoute } from './models/routing/item-route-default';
 import { RedirectToIdComponent } from './containers/redirect-to-id/redirect-to-id.component';
 import { PageNotFoundComponent } from './containers/page-not-found/page-not-found.component';
 import { activityPrefix, itemRouteAsUrlCommand, skillPrefix } from './models/routing/item-route-serialization';
