+++ src/app/store/navigation/current-content/current-content.state.ts
@@ -1,6 +1,6 @@
 import { ContentBreadcrumbs } from 'src/app/models/content/content-breadcrumbs';
 import { ContentRoute } from 'src/app/models/routing/content-route';
-import { appDefaultActivityRoute } from 'src/app/models/routing/item-route';
+import { appDefaultActivityRoute } from 'src/app/models/routing/item-route-default';
 
 export interface State {
   route: ContentRoute|string, // FIXME: the string is current for pages (not by id) which does not match a content route
