+++ src/app/store/navigation/selected-content/selected-content.state.ts
@@ -1,5 +1,6 @@
 import { GroupPage, myGroupsPage, RawGroupRoute } from 'src/app/models/routing/group-route';
-import { appDefaultActivityRoute, appDefaultSkillRoute, ItemRoute } from 'src/app/models/routing/item-route';
+import { ItemRoute } from 'src/app/models/routing/item-route';
+import { appDefaultActivityRoute, appDefaultSkillRoute } from 'src/app/models/routing/item-route-default';
 
 export interface State {
   activity: ItemRoute,
